Be explicit about `sugar.lift` parameter being typed instead of untyped
With Nim's current overloading system, since `lift` has an overload with the first parameter having a type, the first parameter of any `lift` call is assumed to be typed (which is not preferable, discussion about it [here](https://github.com/nim-lang/RFCs/issues/402)). `lift` currently breaks if this behavior is changed, unless we explicitly write `typed`, so do this in case this gets changed.